refactored frontend to use absolute path imports
