+++ client/src/App.js
@@ -2,7 +2,7 @@ import React from 'react';
 import './App.css';
 import { BrowserRouter } from 'react-router-dom';
 
-import Routes from './Routes';
+import Routes from 'src/Routes';
 
 const App = () => {
   return (
