+++ client/src/Routes.js
@@ -1,10 +1,10 @@
 import React from 'react';
 import { Route, Routes as Switch } from 'react-router-dom';
 
-import Layout from './pages/layout';
-import Dashboard from './pages/dashboard';
-import Login from './pages/login';
-import Register from './pages/register';
+import Layout from 'src/pages/layout';
+import Dashboard from 'src/pages/dashboard';
+import Login from 'src/pages/auth/login';
+import Register from 'src/pages/auth/register';
 
 const Routes = () => (
   <Switch>
