+++ client/src/pages/layout/index.js
@@ -1,5 +1,6 @@
 import { Outlet } from 'react-router-dom';
-import Header from './Header';
+
+import Header from 'src/pages/layout/Header';
 
 const Layout = () => {
   return (
