+++ client/src/state/message/actions.js
@@ -1,7 +1,7 @@
 import {
   SET_MESSAGE,
   CLEAR_MESSAGE
-} from '../actionTypes';
+} from 'src/state/actionTypes';
 
 export const setMessage = (message) => ({
   type: SET_MESSAGE,
