+++ client/src/state/store.js
@@ -3,8 +3,8 @@ import { configureStore } from '@reduxjs/toolkit';
 import { composeWithDevTools } from 'redux-devtools-extension';
 import thunk from 'redux-thunk';
 
-import auth from './token/reducer';
-import message from './message/reducer';
+import auth from 'src/state/token/reducer';
+import message from 'src/state/message/reducer';
 
 const middleware = [thunk];
